==> Scripts/02-ValidateLocomotion.Rmd <==
---
title: "02-ValidateLocomotion"
author: "Nicki lentz"
date: "`r Sys.Date()`"
output: html_document
editor_options: 
  chunk_output_type: console
---

# Libraries and data import

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
'%!in%' <- function(x,y)!('%in%'(x,y))

FMPTuesdayNIU2 <- data.table::fread("Data/ProcessedData/FMPTuesdayNIU2.csv") %>% 
  #split movement type to get locomotion category and rename
  mutate(locomotion = sub(' .*', '', movement_type),
         locomotion = case_when(
           locomotion == "Low" ~ "Walking",
           locomotion == "Very" ~ "No locomotion",
           TRUE ~ as.character(locomotion)
         )) #%>% 
  #remove females
  #filter(athlete %!in% c("Yellow-2","Yellow3","Orange-4","Orange-5"))

FMPTuesdayNIU3 <- data.table::fread("Data/ProcessedData/FMPTuesdayNIU3.csv") %>% 
  mutate(locomotion = sub(' .*', '', movement_type),
         locomotion = case_when(
           locomotion == "Low" ~ "Walking",
           locomotion == "Very" ~ "No locomotion",
           TRUE ~ as.character(locomotion)
         )) #%>% 
  #remove females
  #filter(athlete %!in% c("Orange-2","Orange-3"))

#no females in this group
FMPWednesdayNIU1 <- data.table::fread("Data/ProcessedData/FMPWednesdayNIU1.csv") %>% 
  mutate(locomotion = sub(' .*', '', movement_type),
         locomotion = case_when(
           locomotion == "Low" ~ "Walking",
           locomotion == "Very" ~ "No locomotion",
           TRUE ~ as.character(locomotion)
         ))

FMPThursdayNIU2 <- data.table::fread("Data/ProcessedData/FMPThursdayNIU2.csv") %>% 
  mutate(locomotion = sub(' .*', '', movement_type),
         locomotion = case_when(
           locomotion == "Low" ~ "Walking",
           locomotion == "Very" ~ "No locomotion",
           TRUE ~ as.character(locomotion)
         )) #%>% 
  #remove females
  #filter(athlete %!in% c("Yellow-2"))

FMPThursdayNIU3 <- data.table::fread("Data/ProcessedData/FMPThursdayNIU3.csv") %>% 
  mutate(locomotion = sub(' .*', '', movement_type),
         locomotion = case_when(
           locomotion == "Low" ~ "Walking",
           locomotion == "Very" ~ "No locomotion",
           TRUE ~ as.character(locomotion)
         )) #%>% 
  #remove females
  #filter(athlete %!in% c("Orange-2"))
```

```{r}
locomotionFMPtruth <- function(FMPdata, eventIntensity, eventType, athleteID = NULL, groupID, startTime, endTime){
  
  #securing capital first letter
  eventIntensity <- tools::toTitleCase(eventIntensity)
  eventType <- tools::toTitleCase(eventType)
  
  if (is.null(athleteID)) {
    athleteID = FMPdata$athlete
  }
  
  syncFMPData <- FMPdata %>% 
    filter(athlete == athleteID,
           unixTime >= startTime,
           unixTime <= endTime) %>% 
    arrange(athlete, unixTime)
  
  FMPtruth <- syncFMPData %>% 
    mutate(truth = eventType,
           intensity = eventIntensity,
           athlete = paste(groupID, athlete, sep = "_"))
  
  FMPtruth$truth <- factor(FMPtruth$truth, levels = c("Dynamic", "Walking", "Running", "No locomotion"))
  FMPtruth$locomotion <- factor(FMPtruth$locomotion, levels = c("Dynamic", "Walking", "Running", "No locomotion"))
  
  return(FMPtruth)
}

tilePlot <- function(data, drillType, eventType, vel = NULL, dynamicintensity = NULL, nCol = 3){
  #drill type = linear or nonlinear
  # vel = different running velocities performed: 12, 18, 19.8 and 21.6 km/t
  # dynamicintensity = 3 different intensities: cone every 3 sec (medium), all out (high), all out half circle (highhalf)
  if(drillType == "linear"){
    
    if (eventType == "Walking" & is.null(vel)) {
      sub = "Tiles represents the prediction from FMP during Walking at 5.4 km/h"
    } else if (eventType == "Running" & vel == 12){
      sub = "Tiles represents the prediction from FMP during Running at 12 km/h"
    } else if (eventType == "Running" & vel == 18){
      sub = "Tiles represents the prediction from FMP during Running at 18 km/h"
    } else if (eventType == "Running" & vel == 19.8){
      sub = "Tiles represents the prediction from FMP during Running at 19.8 km/h"
    } else if (eventType == "Running" & vel == 21.6){
      sub = "Tiles represents the prediction from FMP during Running at 21.6 km/h"
    }
  } else if (drillType == "nonlinear"){ 
    
    if(eventType == "Dynamic" & dynamicintensity == "medium"){
      sub = "Tiles represents the prediction from FMP during dynamic locomotion in and out of a circle with 
              approximately 3 s between passing each cone through the whole circle"
    } else if (eventType == "Dynamic" & dynamicintensity == "high"){
      sub = "Tiles represents the prediction from FMP during dynamic locomotion in and out of a circle with as much effort as possible
              passing each cone through the whole circle"
    } else if (eventType == "Dynamic" & dynamicintensity == "highhalf"){
      sub = "Tiles represents the prediction from FMP during dynamic locomotion in and out of a circle with as much effort as possible
              passing each cone through half of the circle"
    }
  }
  
  data %>% 
    ggplot() + 
    geom_tile(aes(x = unixTime, y = locomotion, fill = locomotion), 
              colour = "black", width = 0.85, height = 0.85, size = 1) + 
    facet_wrap(~athlete, scale = "free_x", ncol = nCol) + 
    theme_minimal() + 
    theme(legend.position = "top",
          axis.text.y = element_blank(),
          axis.text.x = element_blank(),
          text=element_text(size=35)) + 
    scale_fill_manual("", values=c("No locomotion"="#313131", "Walking"="#0f3698", "Dynamic"="#94281b", "Running"="#3c8001")) + 
    labs(title = "Football movement profile",
         subtitle = sub) + 
    ylab("Football Movement Profile") + 
    xlab("")
}

piePlot <- function(data, title){
  ggthemr::ggthemr('greyscale')
  p1 <- data %>% 
    group_by(athlete) %>% 
    count(locomotion) %>% 
    mutate(percen = (n / sum(n)*100)) %>% 
    ungroup() %>% 
    ggplot(aes(x = "", y = percen, fill=locomotion)) + 
    geom_bar(stat = "identity", width = 1) +
    coord_polar("y", start = 0) + 
    facet_wrap(~athlete) + 
    #geom_text(aes(label = paste0(round(percen, 2), "%")), position = position_stack(vjust = 0.5), size = 12) +
    theme_void(base_size = 20) + 
    scale_fill_manual(values=c("No Locomotion"="#313131", "Walking"="#0f3698", 
                               "Dynamic"="#94281b", "Running"="#3c8001"),
                      labels = c("No Locomotion","Slow Locomotion","Non-linear Locomotion","Linear Locomotion")) + 
    ggtitle(title) + 
    theme(plot.title = element_text(hjust = 0.5, size = 30))
  
  return(p1)
  
}
```

# TuesdayNIU2 Walk

```{r}
Tuesday_NIU2_Walk2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU2",
                                         startTime = as.numeric(lubridate::as_datetime("2022-02-15 08:31:10 UTC")),
                                         endTime = as.numeric(lubridate::as_datetime("2022-02-15 08:31:53 UTC")))

Tuesday_NIU2_Walk3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU2",
                                         startTime = lubridate::as_datetime("2022-02-15 08:32:07 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 08:32:49 UTC"))

Tuesday_NIU2_Walk4 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU2",
                                         startTime = lubridate::as_datetime("2022-02-15 08:33:04 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 08:33:47 UTC"))

Tuesday_NIU2_Walk5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU2",
                                         startTime = lubridate::as_datetime("2022-02-15 08:34:02 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 08:34:43 UTC"))


Tuesday_NIU2_Walk <- rbind(Tuesday_NIU2_Walk2,
                           Tuesday_NIU2_Walk3,
                           Tuesday_NIU2_Walk4,
                           Tuesday_NIU2_Walk5)

Tuesday_NIU2_Walk %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_Walk5, drillType = "linear", eventType = "Walking")
ggsave("Output/Visualization/Tuesday_NIU2_Walk5.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_Walk, "Slow Locomotion (5.4 km/h)")
ggsave("Output/Visualization/Pieplot_Walk.jpeg", height = 12, width = 25)

```

# TuesdayNIU2 Run 10m/3sec

```{r}
Tuesday_NIU2_Run10_2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                           startTime = lubridate::as_datetime("2022-02-15 08:37:37 UTC"),
                                           endTime = lubridate::as_datetime("2022-02-15 08:37:55 UTC"))

Tuesday_NIU2_Run10_3  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:38:10 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:38:28 UTC"))

Tuesday_NIU2_Run10_4  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:38:43 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:39:01 UTC"))

Tuesday_NIU2_Run10_5  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:39:16 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:39:34 UTC"))


Tuesday_NIU2_Run10 <- rbind(Tuesday_NIU2_Run10_2,
                            Tuesday_NIU2_Run10_3,
                            Tuesday_NIU2_Run10_4,
                            Tuesday_NIU2_Run10_5)

Tuesday_NIU2_Run10 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_Run10_5, drillType = "linear", eventType = "Running", vel = 12)
ggsave("Output/Visualization/Tuesday_NIU2_Walk5.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_Run10, "Linear Locomotion (12 km/h)")
ggsave("Output/Visualization/Pieplot_Run10.jpeg", height = 12, width = 25)
```

# TuesdayNIU2 Run 15m/3sec

```{r}
Tuesday_NIU2_Run15_2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                           startTime = lubridate::as_datetime("2022-02-15 08:45:29 UTC"),
                                           endTime = lubridate::as_datetime("2022-02-15 08:45:40 UTC"))

Tuesday_NIU2_Run15_3  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:45:56 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:46:07 UTC"))

Tuesday_NIU2_Run15_4  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:46:23 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:46:34 UTC"))

Tuesday_NIU2_Run15_5  <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "low", eventType = "Running", groupID = "TuesdayNIU2",
                                            startTime = lubridate::as_datetime("2022-02-15 08:46:49 UTC"),
                                            endTime = lubridate::as_datetime("2022-02-15 08:47:00 UTC"))


Tuesday_NIU2_Run15 <- rbind(Tuesday_NIU2_Run15_2,
                            Tuesday_NIU2_Run15_3,
                            Tuesday_NIU2_Run15_4,
                            Tuesday_NIU2_Run15_5)

Tuesday_NIU2_Run15 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_Run10_5, drillType = "linear", eventType = "Running", vel = 18)
ggsave("Output/Visualization/Tuesday_NIU_Walk15_5.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_Run15, "Linear Locomotion (18 km/h)")
ggsave("Output/Visualization/Pieplot_Run15.jpeg", height = 12, width = 25)

```

# TuesdayNIU2 Medium Dynamic
FMPdata, eventIntensity, eventType, athleteID = NULL, groupID, startTime, endTim

```{r}
Tuesday_NIU2_medium_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-2", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:02:40 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:22 UTC")))

Tuesday_NIU2_medium_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-3", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:02:48 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:35 UTC")))

Tuesday_NIU2_medium_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-5", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:02:57 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:39 UTC")))

Tuesday_NIU2_medium_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-6", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:09 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:51 UTC")))

Tuesday_NIU2_medium_Yellow7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-7", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:18 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:57 UTC")))

Tuesday_NIU2_medium_Yellow8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-8", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:27 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:07 UTC")))

Tuesday_NIU2_medium_Yellow9 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-9", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:37 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:17 UTC")))

Tuesday_NIU2_medium_Yellow10 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Yellow-10", groupID = "TuesdayNIU2",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:42 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:28 UTC")))

Tuesday_NIU2_medium_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-2", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:03:55 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:40 UTC")))

Tuesday_NIU2_medium_Orange3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-3", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:01 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:41 UTC")))

Tuesday_NIU2_medium_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-4", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:12 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:57 UTC")))

Tuesday_NIU2_medium_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-5", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:24 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:05:11 UTC")))

Tuesday_NIU2_medium_Orange6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-6", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:34 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:05:14 UTC")))

Tuesday_NIU2_medium_Orange7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-8", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:40 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:05:19 UTC")))

Tuesday_NIU2_medium_Orange8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-8", groupID = "TuesdayNIU2",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:04:49 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:05:26 UTC")))

Tuesday_NIU2_mediumDynamic <- rbind(Tuesday_NIU2_medium_Yellow2, 
                                    Tuesday_NIU2_medium_Yellow3, 
                                    Tuesday_NIU2_medium_Yellow5, 
                                    Tuesday_NIU2_medium_Yellow6, 
                                    Tuesday_NIU2_medium_Yellow7,
                                    Tuesday_NIU2_medium_Yellow8, 
                                    Tuesday_NIU2_medium_Yellow9, 
                                    Tuesday_NIU2_medium_Yellow10, 
                                    Tuesday_NIU2_medium_Orange2, 
                                    Tuesday_NIU2_medium_Orange3,
                                    Tuesday_NIU2_medium_Orange4, 
                                    Tuesday_NIU2_medium_Orange5, 
                                    Tuesday_NIU2_medium_Orange6, 
                                    Tuesday_NIU2_medium_Orange7, 
                                    Tuesday_NIU2_medium_Orange8)
Tuesday_NIU2_mediumDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_mediumDynamic, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "medium")
ggsave("Output/Visualization/Tuesday_NIU2_mediumDynamic.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_mediumDynamic, "Non-linear Locomotion (3 sec per cone)")
ggsave("Output/Visualization/Pieplot_mediumDynamic.jpeg", height = 12, width = 25)
```

# TuesdayNIU2 High Dynamic

```{r}
Tuesday_NIU2_high_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-2", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:06:55 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:07:31 UTC"))

Tuesday_NIU2_high_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-3", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:04 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:07:40 UTC"))

Tuesday_NIU2_high_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-5", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:13 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:07:49"))

Tuesday_NIU2_high_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-6", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:22 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:07:53 UTC"))

Tuesday_NIU2_high_Yellow7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-7", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:30 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:01 UTC"))

Tuesday_NIU2_high_Yellow8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-8", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:38 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:07 UTC"))

Tuesday_NIU2_high_Yellow9 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-9", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:07:46 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:17 UTC"))

Tuesday_NIU2_high_Yellow10 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                 athleteID = "Yellow-10", groupID = "TuesdayNIU2",
                                                 startTime = lubridate::as_datetime("2022-02-15 09:07:53 UTC"),
                                                 endTime = lubridate::as_datetime("2022-02-15 09:08:21 UTC"))

Tuesday_NIU2_high_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-2", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:00 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:33 UTC"))

Tuesday_NIU2_high_Orange3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-3", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:13 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:44 UTC"))

Tuesday_NIU2_high_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-4", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:21 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:08:53 UTC"))

Tuesday_NIU2_high_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-5", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:28 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:09:02 UTC"))

Tuesday_NIU2_high_Orange6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-6", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:36 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:09:03 UTC"))

Tuesday_NIU2_high_Orange7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-7", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:43 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:09:07 UTC"))

Tuesday_NIU2_high_Orange8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-8", groupID = "TuesdayNIU2",
                                                startTime = lubridate::as_datetime("2022-02-15 09:08:49 UTC"),
                                                endTime = lubridate::as_datetime("2022-02-15 09:09:19 UTC"))

Tuesday_NIU2_highDynamic <- rbind(Tuesday_NIU2_high_Yellow2, 
                                  Tuesday_NIU2_high_Yellow3, 
                                  Tuesday_NIU2_high_Yellow5, 
                                  Tuesday_NIU2_high_Yellow6, 
                                  Tuesday_NIU2_high_Yellow7,
                                  Tuesday_NIU2_high_Yellow8,
                                  Tuesday_NIU2_high_Yellow9,
                                  Tuesday_NIU2_high_Yellow10,
                                  Tuesday_NIU2_high_Orange2,
                                  Tuesday_NIU2_high_Orange3,
                                  Tuesday_NIU2_high_Orange4, 
                                  Tuesday_NIU2_high_Orange5, 
                                  Tuesday_NIU2_high_Orange6, 
                                  Tuesday_NIU2_high_Orange7, 
                                  Tuesday_NIU2_high_Orange8)

Tuesday_NIU2_highDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_highDynamic, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "high")
ggsave("Output/Visualization/Tuesday_NIU2_highDynamic.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_highDynamic, "Non-linear Locomotion (all out - full circle)")
ggsave("Output/Visualization/Pieplot_highDynamic.jpeg", height = 12, width = 25)
```

# TuesdayNIU2 High Half Dynamic

```{r}
Tuesday_NIU2_highHalf_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-2", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:19 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:35 UTC")))

Tuesday_NIU2_highHalf_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-3", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:24 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:41 UTC")))

Tuesday_NIU2_highHalf_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-5", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:30 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:47 UTC")))

Tuesday_NIU2_highHalf_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-6", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:36 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:51 UTC")))

Tuesday_NIU2_highHalf_Yellow7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-7", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:43 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:59 UTC")))

Tuesday_NIU2_highHalf_Yellow8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-8", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:10:55 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:10 UTC")))

Tuesday_NIU2_highHalf_Yellow9 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-9", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:01 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:16 UTC")))

Tuesday_NIU2_highHalf_Yellow10 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                     athleteID = "Yellow-10", groupID = "TuesdayNIU2",
                                                     startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:07 UTC")),
                                                     endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:22 UTC")))

Tuesday_NIU2_highHalf_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-2", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:14 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:30 UTC")))

Tuesday_NIU2_highHalf_Orange3 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-3", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:22 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:38 UTC")))

Tuesday_NIU2_highHalf_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-4", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:28 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:44 UTC")))

Tuesday_NIU2_highHalf_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-5", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:35 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:51 UTC")))

Tuesday_NIU2_highHalf_Orange6 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-6", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:42 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:57 UTC")))

Tuesday_NIU2_highHalf_Orange7 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-7", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:48 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:12:03 UTC")))

Tuesday_NIU2_highHalf_Orange8 <- locomotionFMPtruth(FMPTuesdayNIU2, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-8", groupID = "TuesdayNIU2",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 09:11:54 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 09:12:10 UTC")))

Tuesday_NIU2_highHalf <- rbind(Tuesday_NIU2_highHalf_Yellow2, 
                               Tuesday_NIU2_highHalf_Yellow3, 
                               Tuesday_NIU2_highHalf_Yellow5, 
                               Tuesday_NIU2_highHalf_Yellow6, 
                               Tuesday_NIU2_highHalf_Yellow7,
                               Tuesday_NIU2_highHalf_Yellow8, 
                               Tuesday_NIU2_highHalf_Yellow9, 
                               Tuesday_NIU2_highHalf_Yellow10, 
                               Tuesday_NIU2_highHalf_Orange2, 
                               Tuesday_NIU2_highHalf_Orange3,
                               Tuesday_NIU2_highHalf_Orange4,
                               Tuesday_NIU2_highHalf_Orange5, 
                               Tuesday_NIU2_highHalf_Orange6, 
                               Tuesday_NIU2_highHalf_Orange7, 
                               Tuesday_NIU2_highHalf_Orange8)

Tuesday_NIU2_highHalf %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU2_highHalf, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "highhalf")
ggsave("Output/Visualization/Tuesday_NIU2_highHalf.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU2_highHalf, "Non-linear Locomotion (All out - half circle)")
ggsave("Output/Visualization/Pieplot_highHalf.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 Walk

```{r}
Tuesday_NIU3_Walk2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU3",
                                         startTime = lubridate::as_datetime("2022-02-15 10:20:10 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 10:20:50 UTC"))

Tuesday_NIU3_Walk3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU3",
                                         startTime = lubridate::as_datetime("2022-02-15 10:21:03 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 10:21:43 UTC"))

Tuesday_NIU3_Walk4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU3",
                                         startTime = lubridate::as_datetime("2022-02-15 10:21:57 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 10:22:37 UTC"))

Tuesday_NIU3_Walk5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "low", eventType = "Walking", groupID = "TuesdayNIU3",
                                         startTime = lubridate::as_datetime("2022-02-15 10:22:50 UTC"),
                                         endTime = lubridate::as_datetime("2022-02-15 10:23:30 UTC"))


Tuesday_NIU3_Walk <- rbind(Tuesday_NIU3_Walk2,
                           Tuesday_NIU3_Walk3,
                           Tuesday_NIU3_Walk4,
                           Tuesday_NIU3_Walk5)

Tuesday_NIU3_Walk %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_Walk2, drillType = "linear", eventType = "Walking")
ggsave("Output/Visualization/Tuesday_NIU3_Walk2.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_Walk, "Slow Locomotion (5.4 km/h)")
ggsave("Output/Visualization/Pieplot_Walk_NIU3.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 Run 10m/3sec

```{r}
Tuesday_NIU3_Run10_3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Running", groupID = "TuesdayNIU3",
                                           startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:27:12 UTC")),
                                           endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:27:30 UTC")))

Tuesday_NIU3_Run10_4  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:27:42 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:28:00 UTC")))

Tuesday_NIU3_Run10_5  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:28:12 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:28:30 UTC")))

Tuesday_NIU3_Run10_6  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:28:43 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:29:01 UTC")))


Tuesday_NIU3_Run10 <- rbind(Tuesday_NIU3_Run10_3,
                            Tuesday_NIU3_Run10_4,
                            Tuesday_NIU3_Run10_5,
                            Tuesday_NIU3_Run10_6)

Tuesday_NIU3_Run10 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_Run10_4, drillType = "linear", eventType = "Running", vel = 12)
ggsave("Output/Visualization/Tuesday_NIU3_Run10_4.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_Run10, "Linear Locomotion (12 km/h)")
ggsave("Output/Visualization/Pieplot_Run10_NIU3.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 Run 15m/3sec

```{r}
Tuesday_NIU3_Run15_2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "TuesdayNIU3",
                                           startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:39:35 UTC")),
                                           endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:39:41 UTC")))

Tuesday_NIU3_Run15_3  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:40:02 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:40:08 UTC")))

Tuesday_NIU3_Run15_4  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:40:29 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:40:35 UTC")))

Tuesday_NIU3_Run15_5  <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "TuesdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:40:58 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:41:04 UTC")))


Tuesday_NIU3_Run15 <- rbind(Tuesday_NIU3_Run15_2,
                            Tuesday_NIU3_Run15_3,
                            Tuesday_NIU3_Run15_4,
                            Tuesday_NIU3_Run15_5)

Tuesday_NIU3_Run15 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_Run15_3, drillType = "linear", eventType = "Running", vel = 18)
ggsave("Output/Visualization/Tuesday_NIU3_Run15_3.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_Run15, "Linear Locomotion (18 km/h)")
ggsave("Output/Visualization/Pieplot_Run15_NIU3.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 Medium Dynamic 1

```{r}
Tuesday_NIU3_medium_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-2", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:48:35 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:20 UTC")))

Tuesday_NIU3_medium_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-4", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:48:49 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:34 UTC")))

Tuesday_NIU3_medium_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-5", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:01 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:47 UTC")))

Tuesday_NIU3_medium_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-2", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:17 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:50:04 UTC")))

Tuesday_NIU3_medium_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-3", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:31 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:50:19 UTC")))

Tuesday_NIU3_medium_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-5", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:49:46 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:50:31 UTC")))

Tuesday_NIU3_medium_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-6", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:50:02 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:50:48 UTC")))



Tuesday_NIU3_mediumDynamic <- rbind(Tuesday_NIU3_medium_Orange2, 
                                    Tuesday_NIU3_medium_Orange4, 
                                    Tuesday_NIU3_medium_Orange5,
                                    Tuesday_NIU3_medium_Yellow2, 
                                    Tuesday_NIU3_medium_Yellow3, 
                                    Tuesday_NIU3_medium_Yellow5, 
                                    Tuesday_NIU3_medium_Yellow6)

Tuesday_NIU3_mediumDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_mediumDynamic, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "medium")
ggsave("Output/Visualization/Tuesday_NIU3_mediumDynamic.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_mediumDynamic, "Non-linear Locomotion (3 s per cone)")
ggsave("Output/Visualization/Pieplot_mediumDynamic_NIU3.jpeg", height = 12, width = 25)

```

# TuesdayNIU3 Medium Dynamic 2

```{r}
Tuesday_NIU3_medium2_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Orange-2", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:52:11 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:52:57 UTC")))

Tuesday_NIU3_medium2_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Orange-4", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:52:26 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:11 UTC")))

Tuesday_NIU3_medium2_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Orange-5", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:52:37 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:22 UTC")))

Tuesday_NIU3_medium2_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Yellow-2", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:52:53 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:40 UTC")))

Tuesday_NIU3_medium2_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Yellow-3", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:08 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:54 UTC")))

Tuesday_NIU3_medium2_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Yellow-5", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:20 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:54:05 UTC")))

Tuesday_NIU3_medium2_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                   athleteID = "Yellow-6", groupID = "TuesdayNIU3",
                                                   startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:53:37 UTC")),
                                                   endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:54:23 UTC")))


Tuesday_NIU3_mediumDynamic2 <- rbind(Tuesday_NIU3_medium2_Orange2, 
                                     Tuesday_NIU3_medium2_Orange4, 
                                     Tuesday_NIU3_medium2_Orange5,
                                     Tuesday_NIU3_medium2_Yellow2, 
                                     Tuesday_NIU3_medium2_Yellow3, 
                                     Tuesday_NIU3_medium2_Yellow5, 
                                     Tuesday_NIU3_medium2_Yellow6)

Tuesday_NIU3_mediumDynamic2 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_mediumDynamic2, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "medium")
ggsave("Output/Visualization/Tuesday_NIU3_mediumDynamic2.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_mediumDynamic2, "Non-linear Locomotion (3 s per cone)")
ggsave("Output/Visualization/Pieplot_mediumDynamic2_NIU3.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 Medium Dynamic (No mid cones)

```{r}
Tuesday_NIU3_noCone_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-2", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:55:10 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:55:56 UTC")))

Tuesday_NIU3_noCone_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-4", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:55:24 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:10 UTC")))

Tuesday_NIU3_noCone_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Orange-5", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:55:36 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:23 UTC")))

Tuesday_NIU3_noCone_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-2", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:55:49 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:36 UTC")))

Tuesday_NIU3_noCone_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-3", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:04 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:51 UTC")))

Tuesday_NIU3_noCone_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-5", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:18 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:57:04 UTC")))

Tuesday_NIU3_noCone_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "medium", eventType = "Dynamic",
                                                  athleteID = "Yellow-6", groupID = "TuesdayNIU3",
                                                  startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:56:33 UTC")),
                                                  endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:57:19 UTC")))



Tuesday_NIU3_noCone <- rbind(Tuesday_NIU3_noCone_Orange2, 
                             Tuesday_NIU3_noCone_Orange4, 
                             Tuesday_NIU3_noCone_Orange5,
                             Tuesday_NIU3_noCone_Yellow2, 
                             Tuesday_NIU3_noCone_Yellow3, 
                             Tuesday_NIU3_noCone_Yellow5, 
                             Tuesday_NIU3_noCone_Yellow6)

Tuesday_NIU3_noCone %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_noCone, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "medium")
ggsave("Output/Visualization/Tuesday_NIU3_noCone.jpeg", width = 25, height = 10)


```

# TuesdayNIU3 High Dynamic

```{r}
Tuesday_NIU3_high_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-2", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:58:33 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:12 UTC")))

Tuesday_NIU3_high_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-4", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:58:53 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:24 UTC")))

Tuesday_NIU3_high_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Orange-5", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:05 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:37 UTC")))

Tuesday_NIU3_high_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-2", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:17 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:52 UTC")))

Tuesday_NIU3_high_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-3", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:32 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:00:10 UTC")))

Tuesday_NIU3_high_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-5", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 10:59:47 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:00:19 UTC")))

Tuesday_NIU3_high_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "high", eventType = "Dynamic",
                                                athleteID = "Yellow-6", groupID = "TuesdayNIU3",
                                                startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:00:03 UTC")),
                                                endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:00:33 UTC")))



Tuesday_NIU3_highDynamic <- rbind(Tuesday_NIU3_high_Orange2, 
                                  Tuesday_NIU3_high_Orange4, 
                                  Tuesday_NIU3_high_Orange5,
                                  Tuesday_NIU3_high_Yellow2, 
                                  Tuesday_NIU3_high_Yellow3, 
                                  Tuesday_NIU3_high_Yellow5, 
                                  Tuesday_NIU3_high_Yellow6)

Tuesday_NIU3_highDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_highDynamic, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "high")
ggsave("Output/Visualization/Tuesday_NIU3_highDynamic.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_highDynamic, "Non-linear Locomotion (All out - full circle)")
ggsave("Output/Visualization/Pieplot_highDynamic_NIU3.jpeg", height = 12, width = 25)
```

# TuesdayNIU3 High Half Dynamic

```{r}
Tuesday_NIU3_highHalf_Orange2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-2", groupID = "TuesdayNIU3", 
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:06:14 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:06:32 UTC")))

Tuesday_NIU3_highHalf_Orange4 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-4", groupID = "TuesdayNIU3", 
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:06:52 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:07:07 UTC")))

Tuesday_NIU3_highHalf_Orange5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Orange-5", groupID = "TuesdayNIU3", 
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:07:23 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:07:39 UTC")))

Tuesday_NIU3_highHalf_Yellow2 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-2", groupID = "TuesdayNIU3", 
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:07:51 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:08:07 UTC")))

Tuesday_NIU3_highHalf_Yellow3 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-3", groupID = "TuesdayNIU3",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:08:25 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:08:42 UTC")))

Tuesday_NIU3_highHalf_Yellow5 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-5", groupID = "TuesdayNIU3",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:09:05 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:09:19 UTC")))

Tuesday_NIU3_highHalf_Yellow6 <- locomotionFMPtruth(FMPTuesdayNIU3, eventIntensity = "highhalf", eventType = "Dynamic",
                                                    athleteID = "Yellow-6", groupID = "TuesdayNIU3",
                                                    startTime = as.numeric(lubridate::as_datetime("2022-02-15 11:09:41 UTC")),
                                                    endTime = as.numeric(lubridate::as_datetime("2022-02-15 11:09:56 UTC")))



Tuesday_NIU3_highHalf <- rbind(Tuesday_NIU3_highHalf_Orange2, 
                               Tuesday_NIU3_highHalf_Orange4, 
                               Tuesday_NIU3_highHalf_Orange5,
                               Tuesday_NIU3_highHalf_Yellow2, 
                               Tuesday_NIU3_highHalf_Yellow3, 
                               Tuesday_NIU3_highHalf_Yellow5, 
                               Tuesday_NIU3_highHalf_Yellow6)

Tuesday_NIU3_highHalf %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Tuesday_NIU3_highHalf, drillType = "nonlinear", eventType = "Dynamic", dynamicintensity = "highhalf")
ggsave("Output/Visualization/Tuesday_NIU3_highHalf.jpeg", width = 25, height = 10)

plot <- piePlot(Tuesday_NIU3_highHalf, "Non-linear Locomotion (All out - half circle)")
ggsave("Output/Visualization/Pieplot_highhalf_NIU3.jpeg", height = 12, width = 25)
```

# WednesdayNIU1 Run 10m/3sec

```{r}
Wednesday_NIU1_Run10_1 <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "medium", eventType = "Running", groupID = "WednesdayNIU1",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:32:34 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:32:46 UTC")))

Wednesday_NIU1_Run10_2  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "medium", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:33:04 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:33:16 UTC")))

Wednesday_NIU1_Run10_3  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "medium", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:33:32 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:33:44 UTC")))

Wednesday_NIU1_Run10_4  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "medium", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:33:58 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:34:10 UTC")))


Wednesday_NIU1_Run10 <- rbind(Wednesday_NIU1_Run10_1,
                              Wednesday_NIU1_Run10_2,
                              Wednesday_NIU1_Run10_3,
                              Wednesday_NIU1_Run10_4)

Wednesday_NIU1_Run10 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Wednesday_NIU1_Run10, drillType = "linear", eventType = "Running", vel = 12)
ggsave("Output/Visualization/Wednesday_NIU1_Run10.jpeg", width = 25, height = 10)

plot <- piePlot(Wednesday_NIU1_Run10, "Linear Locomotion (12 km/h)")
ggsave("Output/Visualization/Pieplot_Run10_NIU1.jpeg", height = 12, width = 25)
```

# WednesdayNIU1 Run 15m/3sec

```{r}
Wednesday_NIU1_Run15_1 <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:40:30 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:40:36 UTC")))

Wednesday_NIU1_Run15_2  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:40:54 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:41:00 UTC")))

Wednesday_NIU1_Run15_3  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:41:21 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:41:27 UTC")))

Wednesday_NIU1_Run15_4  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:41:45 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:41:51 UTC")))


Wednesday_NIU1_Run15 <- rbind(Wednesday_NIU1_Run15_1,
                              Wednesday_NIU1_Run15_2,
                              Wednesday_NIU1_Run15_3,
                              Wednesday_NIU1_Run15_4)

Wednesday_NIU1_Run15 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Wednesday_NIU1_Run15, drillType = "linear", eventType = "Running", vel = 18)
ggsave("Output/Visualization/Wednesday_NIU1_Run15.jpeg", width = 25, height = 10)

plot <- piePlot(Wednesday_NIU1_Run15, "Linear Locomotion (18 km/h)")
ggsave("Output/Visualization/Pieplot_Run15_NIU1.jpeg", height = 12, width = 25)
```

# WednesdayNIU1 Run 16.5m / 3 sec

```{r}
Wednesday_NIU1_Run16_1 <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:47:09 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:47:15 UTC")))

Wednesday_NIU1_Run16_2  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:47:34 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:47:40 UTC")))

Wednesday_NIU1_Run16_3  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:48:01 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:48:07 UTC")))

Wednesday_NIU1_Run16_4  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:48:28 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:48:34 UTC")))


Wednesday_NIU1_Run16 <- rbind(Wednesday_NIU1_Run16_1,
                              Wednesday_NIU1_Run16_2,
                              Wednesday_NIU1_Run16_3,
                              Wednesday_NIU1_Run16_4)

Wednesday_NIU1_Run16 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Wednesday_NIU1_Run16, drillType = "linear", eventType = "Running", vel = 19.8)
ggsave("Output/Visualization/Wednesday_NIU1_Run16.jpeg", width = 25, height = 10)

plot <- piePlot(Wednesday_NIU1_Run16, "Linear Locomotion (19.8 km/h)")
ggsave("Output/Visualization/Pieplot_Run16_NIU1.jpeg", height = 12, width = 25)
```

# WednesdayNIU 1 Run 18m / 3 sec

```{r}
Wednesday_NIU1_Run18_1 <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:54:27 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:54:33 UTC")))

Wednesday_NIU1_Run18_2  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:54:58 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:55:04 UTC")))

Wednesday_NIU1_Run18_3  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:55:30 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:55:36 UTC")))

Wednesday_NIU1_Run18_4  <- locomotionFMPtruth(FMPWednesdayNIU1, eventIntensity = "high", eventType = "Running", groupID = "WednesdayNIU1",
                                              startTime = as.numeric(lubridate::as_datetime("2022-02-16 08:56:02 UTC")),
                                              endTime = as.numeric(lubridate::as_datetime("2022-02-16 08:56:08 UTC")))


Wednesday_NIU1_Run18 <- rbind(Wednesday_NIU1_Run18_1,
                              Wednesday_NIU1_Run18_2,
                              Wednesday_NIU1_Run18_3,
                              Wednesday_NIU1_Run18_4)

Wednesday_NIU1_Run18 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Wednesday_NIU1_Run18, drillType = "linear", eventType = "Running", vel = 21.6)
ggsave("Output/Visualization/Wednesday_NIU1_Run18.jpeg", width = 25, height = 10)

plot <- piePlot(Wednesday_NIU1_Run18, "Linear Locomotion (21.6 km/h)")
ggsave("Output/Visualization/Pieplot_Run18_NIU1.jpeg", height = 12, width = 25)
```

# ThursdayNIU2 Run 15m / 3 sec

```{r}
Thursday_NIU2_Run15_1 <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:26:07 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:26:13 UTC")))

Thursday_NIU2_Run15_2  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:26:35 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:26:41 UTC")))

Thursday_NIU2_Run15_3  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:27:01 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:27:07 UTC")))

Thursday_NIU2_Run15 <- rbind(Thursday_NIU2_Run15_1,
                             Thursday_NIU2_Run15_2,
                             Thursday_NIU2_Run15_3)
Thursday_NIU2_Run15 %>%
  select(truth, locomotion) %>%
  irr::agree(.)

tilePlot(Thursday_NIU2_Run15_1, drillType = "linear", eventType = "Running", vel = 18)
ggsave("Output/Visualization/Thursday_NIU2_Run15_1.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU2_Run15, "Linear Locomotion (18 km/h)")
ggsave("Output/Visualization/Pieplot_Run18_NIU2.jpeg", height = 12, width = 25)
```



# ThursdayNIU2 Run 16.5m / 3sec

```{r}
Thursday_NIU2_Run16_1 <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:29:39 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:29:45 UTC")))

Thursday_NIU2_Run16_2  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:30:18 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:30:24 UTC")))

Thursday_NIU2_Run16_3  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:30:52 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:30:58 UTC")))

Thursday_NIU2_Run16 <- rbind(Thursday_NIU2_Run16_1,
                             Thursday_NIU2_Run16_2,
                             Thursday_NIU2_Run16_3)
Thursday_NIU2_Run16 %>%
  select(truth, locomotion) %>%
  irr::agree(.)

tilePlot(Thursday_NIU2_Run16, drillType = "linear", eventType = "Running", vel = 19.8)
ggsave("Output/Visualization/Thursday_NIU2_Run16.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU2_Run16, "Linear Locomotion (19.8 km/h)")
ggsave("Output/Visualization/Pieplot_Run16_NIU2.jpeg", height = 12, width = 25)
```

# ThursdayNIU2 Run 18 m / 3 sec

```{r}
Thursday_NIU2_Run18_1 <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:32:55 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:33:01 UTC")))

Thursday_NIU2_Run18_2  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:33:31 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:33:37 UTC")))

Thursday_NIU2_Run18_3  <- locomotionFMPtruth(FMPThursdayNIU2, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU2",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 08:34:07 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 08:34:13 UTC")))

Thursday_NIU2_Run18 <- rbind(Thursday_NIU2_Run18_1,
                             Thursday_NIU2_Run18_2,
                             Thursday_NIU2_Run18_3)
Thursday_NIU2_Run18 %>%
  select(truth, locomotion) %>%
  irr::agree(.)

tilePlot(Thursday_NIU2_Run18, drillType = "linear", eventType = "Running", vel = 21.6)
ggsave("Output/Visualization/Thursday_NIU2_Run18.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU2_Run18, "Linear Locomotion (21.6 km/h)")
ggsave("Output/Visualization/Pieplot_Run18_NIU2.jpeg", height = 12, width = 25)
```

# ThursdayNIU3 Run 15m / 3 sec

```{r}
Thursday_NIU3_Run15_1 <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:10:16 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:10:22 UTC")))

Thursday_NIU3_Run15_2  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:10:49 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:10:55 UTC")))

Thursday_NIU3_Run15_3  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:11:18 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:11:24 UTC")))


Thursday_NIU3_Run15 <- rbind(Thursday_NIU3_Run15_1,
                             Thursday_NIU3_Run15_2,
                             Thursday_NIU3_Run15_3)

Thursday_NIU3_Run15 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Thursday_NIU3_Run15, drillType = "linear", eventType = "Running", vel = 18)
ggsave("Output/Visualization/Thursday_NIU3_Run15.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU3_Run15, "Linear Locomotion (18 km/h)")
ggsave("Output/Visualization/Pieplot_Run15_NIU3.jpeg", height = 12, width = 25)
```

# ThursdayNIU3 Run 16.5 / 3 sec

```{r}
Thursday_NIU3_Run16_1 <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:13:48 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:13:54 UTC")))

Thursday_NIU3_Run16_2  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:14:25 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:14:31 UTC")))

Thursday_NIU3_Run16_3  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:15:10 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:15:16 UTC")))


Thursday_NIU3_Run16 <- rbind(Thursday_NIU3_Run16_1,
                             Thursday_NIU3_Run16_2,
                             Thursday_NIU3_Run16_3)

Thursday_NIU3_Run16 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Thursday_NIU3_Run16, drillType = "linear", eventType = "Running", vel = 19.8)
ggsave("Output/Visualization/Thursday_NIU3_Run16.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU3_Run16, "Linear Locomotion (19.8 km/h)")
ggsave("Output/Visualization/Pieplot_Run16_NIU3.jpeg", height = 12, width = 25)
```

# ThursdayNIU3 Run 18 / 3sec

```{r}
Thursday_NIU3_Run18_1 <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                            startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:17:31 UTC")),
                                            endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:17:37 UTC")))

Thursday_NIU3_Run18_2  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:18:10 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:18:16 UTC")))

Thursday_NIU3_Run18_3  <- locomotionFMPtruth(FMPThursdayNIU3, eventIntensity = "high", eventType = "Running", groupID = "ThursdayNIU3",
                                             startTime = as.numeric(lubridate::as_datetime("2022-02-17 11:18:53 UTC")),
                                             endTime = as.numeric(lubridate::as_datetime("2022-02-17 11:18:59 UTC")))


Thursday_NIU3_Run18 <- rbind(Thursday_NIU3_Run18_1,
                             Thursday_NIU3_Run18_2,
                             Thursday_NIU3_Run18_3)

Thursday_NIU3_Run18 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

tilePlot(Thursday_NIU3_Run18, drillType = "linear", eventType = "Running", vel = 21.6)
ggsave("Output/Visualization/Thursday_NIU3_Run18.jpeg", width = 25, height = 10)

plot <- piePlot(Thursday_NIU3_Run18, "Linear Locomotion (21.6 km/h)")
ggsave("Output/Visualization/Pieplot_Run18_NIU3.jpeg", height = 12, width = 25)
```

# All Walk

```{r}
Walk <- rbind(Tuesday_NIU2_Walk,
              Tuesday_NIU3_Walk)

Walk_perc <- Walk %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

WalkAgree <- Walk %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

Run10 <- rbind(Tuesday_NIU2_Run10,
               Tuesday_NIU3_Run10,
               Wednesday_NIU1_Run10)

Run10_perc <- Run10 %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

Run10Agree <- Run10 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

Run15 <- rbind(Tuesday_NIU2_Run15,
               Tuesday_NIU3_Run15,
               Wednesday_NIU1_Run15,
               Thursday_NIU2_Run15,
               Thursday_NIU3_Run15)

Run15_perc <- Run15 %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

Run15Agree <- Run15 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

Run16 <- rbind(Wednesday_NIU1_Run16,
               Thursday_NIU2_Run16,
               Thursday_NIU3_Run16)

Run16_perc <- Run16 %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

Run16Agree <- Run16 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

Run18 <- rbind(Wednesday_NIU1_Run18,
               Thursday_NIU2_Run18,
               Thursday_NIU3_Run18)

Run18_perc <- Run18 %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

Run18Agree <- Run18 %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

mediumDynamic <- rbind(Tuesday_NIU2_mediumDynamic,
                       Tuesday_NIU3_mediumDynamic,
                       Tuesday_NIU3_mediumDynamic2)

mediumDynamic_perc <- mediumDynamic %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

mediumDynamicAgree <- mediumDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

highDynamic <- rbind(Tuesday_NIU2_highDynamic,
                     Tuesday_NIU3_highDynamic)

highDynamic_perc <- highDynamic %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

highDynamicAgree <- highDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)

highhalfDynamic <- rbind(Tuesday_NIU2_highHalf,
                         Tuesday_NIU3_highHalf)

highhalfDynamic_perc <- highhalfDynamic %>% 
  count(locomotion) %>% 
  mutate(perc = n/sum(n)*100)

highhalfAgree <- highhalfDynamic %>% 
  select(truth, locomotion) %>% 
  irr::agree(.)
```

```{r}
#Table for percent distribution for failed predictions

percentDisagreement <- 
  data.frame(
    event = c("Slow","Linear","Linear","Linear","Linear","Non-linear","Non-linear","Non-linear"),
    velocity = c("5.4","12","18","19.6","21.6","3 sec per cone","All out (full circle)","All out (half circle)"),
    NoLocomotion = c(0.3830369, 0L, 0L, 0L, 0L, 0.1571092, 0L, 0L),
    SlowLocomotion = c(99.2065663, 0L, 0L, 0L, 0L, 6.2843676, 0L, 0L),
    LinearLocomotion = c(0, 95.486111, 88.31169, 81.00358, 74.01434, 17.8318932, 1.699717, 2.017291),
    NonLinearLocomotion = c(0.4103967, 4.513889, 11.68831, 18.99642, 25.98566, 75.7266300, 98.300283, 97.982709)
    
  
)

data.table::fwrite(percentDisagreement, "Data/ProcessedData/percentDisagreement.csv")

```


```{r}
totalAgreement <- 
  data.frame(
    event = c("Slow","Linear","Linear","Linear","Linear","Non-linear","Non-linear","Non-linear"),
    velocity = c("5.4","12","18","19.6","21.6","3 sec per cone","All out (full circle)","All out (half circle)"),
    total_time = c(WalkAgree$subjects, Run10Agree$subjects, Run15Agree$subjects,
                   Run16Agree$subjects, Run18Agree$subjects, mediumDynamicAgree$subjects,
                   highDynamicAgree$subjects, highhalfAgree$subjects),
    agreement = c(WalkAgree$value, Run10Agree$value, Run15Agree$value,
                  Run16Agree$value, Run18Agree$value, mediumDynamicAgree$value,
                  highDynamicAgree$value, highhalfAgree$value)
  )

data.table::fwrite(totalAgreement, "Data/ProcessedData/totalAgreement.csv")
```

# Difference between each group

```{r}
Tuesday_NIU2_Walk_agree <- 
  Tuesday_NIU2_Walk %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_Walk_agree <- 
  Tuesday_NIU3_Walk %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU2_Run10_agree <- 
  Tuesday_NIU2_Run10 %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_Run10_agree <- 
  Tuesday_NIU3_Run10 %>% select(truth, locomotion) %>% irr::agree(.)
Wednesday_NIU1_Run10_agree <- 
  Wednesday_NIU1_Run10 %>% select(truth, locomotion) %>% irr::agree(.)

Tuesday_NIU2_Run15_agree <- 
  Tuesday_NIU2_Run15 %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_Run15_agree <- 
  Tuesday_NIU3_Run15 %>% select(truth, locomotion) %>% irr::agree(.)
Wednesday_NIU1_Run15_agree <- 
  Wednesday_NIU1_Run15 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU2_Run15_agree <- 
  Thursday_NIU2_Run15 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU3_Run15_agree <- 
  Thursday_NIU3_Run15 %>% select(truth, locomotion) %>% irr::agree(.)

Wednesday_NIU1_Run16_agree <- 
  Wednesday_NIU1_Run16 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU2_Run16_agree <- 
  Thursday_NIU2_Run16 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU3_Run16_agree <- 
  Thursday_NIU3_Run16 %>% select(truth, locomotion) %>% irr::agree(.)

Wednesday_NIU1_Run18_agree <- 
  Wednesday_NIU1_Run18 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU2_Run18_agree <- 
  Thursday_NIU2_Run18 %>% select(truth, locomotion) %>% irr::agree(.)
Thursday_NIU3_Run18_agree <- 
  Thursday_NIU3_Run18 %>% select(truth, locomotion) %>% irr::agree(.)

Tuesday_NIU2_mediumDynamic_agree <- 
  Tuesday_NIU2_mediumDynamic %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_mediumDynamic_agree <- 
  Tuesday_NIU3_mediumDynamic %>% select(truth, locomotion) %>% irr::agree(.)

Tuesday_NIU2_highDynamic_agree <- 
  Tuesday_NIU2_highDynamic %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_highDynamic_agree <- 
  Tuesday_NIU3_highDynamic %>% select(truth, locomotion) %>% irr::agree(.)

Tuesday_NIU2_highHalf_agree <- 
  Tuesday_NIU2_highHalf %>% select(truth, locomotion) %>% irr::agree(.)
Tuesday_NIU3_highHalf_agree <- 
  Tuesday_NIU3_highHalf %>% select(truth, locomotion) %>% irr::agree(.)

differenceAgreement <- data.frame(
  event = c("Slow","Linear","Linear","Linear","Linear","Non-linear","Non-linear","Non-linear"),
  velocity = c("5.4","12","18","19.6","21.6","3 sec per cone","All out (full circle)","All out (half circle)"),
  NIU1 = c(NA, Wednesday_NIU1_Run10_agree$value,
           Wednesday_NIU1_Run15_agree$value, Wednesday_NIU1_Run16_agree$value,
           Wednesday_NIU1_Run18_agree$value, NA, NA, NA),
  
  NIU2 = c(Tuesday_NIU2_Walk_agree$value, Tuesday_NIU2_Run10_agree$value,
           Tuesday_NIU2_Run15_agree$value, Thursday_NIU2_Run16_agree$value,
           Thursday_NIU2_Run18_agree$value, Tuesday_NIU2_mediumDynamic_agree$value,
           Tuesday_NIU2_highDynamic_agree$value, Tuesday_NIU2_highHalf_agree$value),
  
  NIU3 = c(Tuesday_NIU3_Walk_agree$value, Tuesday_NIU3_Run10_agree$value,
           Tuesday_NIU3_Run15_agree$value, Thursday_NIU3_Run16_agree$value,
           Thursday_NIU3_Run18_agree$value, Tuesday_NIU3_mediumDynamic_agree$value,
           Tuesday_NIU3_highDynamic_agree$value, Tuesday_NIU3_highHalf_agree$value),
  
  mean = c(WalkAgree$value, Run10Agree$value, Run15Agree$value,
                  Run16Agree$value, Run18Agree$value, mediumDynamicAgree$value,
                  highDynamicAgree$value, highhalfAgree$value)
)


#differenceAgreement$Mean <- rowMeans(differenceAgreement[ ,3:5], na.rm=TRUE)
data.table::fwrite(differenceAgreement, "Data/ProcessedData/differenceAgreement.csv")

```

# PIE PLOTS

```{r}
FMPThursdayNIU2_freq <- FMPThursdayNIU2 %>% 
  group_by(athlete, drillType) %>% 
  count(newFMP) %>%
  mutate(percen = n / sum(n)) %>% 
  ungroup()



p2 <- Tuesday_NIU2_Walk %>% 
  group_by(athlete) %>% 
  count(locomotion) %>% 
  mutate(percen = (n / sum(n)*100)) %>% 
  ungroup() %>% 
  ggplot(aes(x = "", y = percen, fill=locomotion)) + 
  geom_bar(stat = "identity", width = 1) +
  coord_polar("y", start = 0) + 
  facet_wrap(~athlete) + 
  geom_text(aes(label = paste0(round(percen, 2), "%")), position = position_stack(vjust = 0.5)) +
  theme_void() + 
  scale_fill_manual(values=c("No Locomotion"="#313131", "Walking"="#0f3698", 
                             "Dynamic"="#94281b", "Running"="#3c8001"),
                    labels = c("No Locomotion","Slow Locomotion","Non-linear Locomotion","Linear Locomotion")) + 
  ggtitle("Test") + 
  theme(plot.title = element_text(hjust = 0.5, size = 30))
```

# SCATTER PIE PLOT

```{r}
Tuesday_NIU2_Walk <- Tuesday_NIU2_Walk %>% mutate(drill = "Slow")
Tuesday_NIU2_Run10 <- Tuesday_NIU2_Run10 %>% mutate(drill = "Linear10")
Tuesday_NIU2_Run15 <- Tuesday_NIU2_Run15 %>% mutate(drill = "Linear15")
Thursday_NIU2_Run15 <- Thursday_NIU2_Run15 %>% mutate(drill = "Linear15")
Thursday_NIU2_Run16 <- Thursday_NIU2_Run16 %>% mutate(drill = "Linear16")
Thursday_NIU2_Run18 <- Thursday_NIU2_Run18 %>% mutate(drill = "Linear18") 
Tuesday_NIU2_mediumDynamic <- Tuesday_NIU2_mediumDynamic %>% mutate(drill = "NonLinearSlow")
Tuesday_NIU2_highDynamic <- Tuesday_NIU2_highDynamic %>% mutate(drill = "NonLinearFast")
Tuesday_NIU2_highHalf <- Tuesday_NIU2_highHalf %>% mutate(drill = "NonLinearHalf")

Tuesday_NIU3_Walk <- Tuesday_NIU3_Walk %>% mutate(drill = "Slow")
Tuesday_NIU3_Run10 <- Tuesday_NIU3_Run10 %>% mutate(drill = "Linear10")
Tuesday_NIU3_Run15 <- Tuesday_NIU3_Run15 %>% mutate(drill = "Linear15")
Thursday_NIU3_Run15 <- Thursday_NIU3_Run15 %>% mutate(drill = "Linear15")
Thursday_NIU3_Run16 <- Thursday_NIU3_Run16 %>% mutate(drill = "Linear16")
Thursday_NIU3_Run18 <- Thursday_NIU3_Run18 %>% mutate(drill = "Linear18") 
Tuesday_NIU3_mediumDynamic <- Tuesday_NIU3_mediumDynamic %>% mutate(drill = "NonLinearSlow")
Tuesday_NIU3_highDynamic <- Tuesday_NIU3_highDynamic %>% mutate(drill = "NonLinearFast")
Tuesday_NIU3_highHalf <- Tuesday_NIU3_highHalf %>% mutate(drill = "NonLinearHalf")

Wednesday_NIU1_Run10 <- Wednesday_NIU1_Run10 %>% mutate(drill = "Linear10")
Wednesday_NIU1_Run15 <- Wednesday_NIU1_Run15 %>% mutate(drill = "Linear15")
Wednesday_NIU1_Run16 <- Wednesday_NIU1_Run16 %>% mutate(drill = "Linear16")
Wednesday_NIU1_Run18 <- Wednesday_NIU1_Run18 %>% mutate(drill = "Linear18")

controlledDrills <- rbind(Tuesday_NIU2_Walk, Tuesday_NIU2_Run10, Tuesday_NIU2_Run15, Thursday_NIU2_Run15, 
            Thursday_NIU2_Run16, Thursday_NIU2_Run18, Tuesday_NIU2_mediumDynamic, Tuesday_NIU2_highDynamic,
            Tuesday_NIU2_highHalf, Tuesday_NIU3_Walk, Tuesday_NIU3_Run10, Tuesday_NIU3_Run15, 
            Thursday_NIU3_Run15, Thursday_NIU3_Run16, Thursday_NIU3_Run18, Tuesday_NIU3_mediumDynamic,
            Tuesday_NIU3_highDynamic, Tuesday_NIU3_highHalf, Wednesday_NIU1_Run10, Wednesday_NIU1_Run15,
            Wednesday_NIU1_Run16, Wednesday_NIU1_Run18)
```


```{r}
t <- controlledDrills %>% 
  group_by(athlete, drill) %>% 
  count(locomotion) %>% 
  mutate(percen = (n / sum(n)*100)) %>% 
  ungroup() %>% 
  pivot_wider(id_cols = c(athlete, drill), names_from = locomotion, values_from = percen, values_fill = 0L) %>%
  mutate(
    #Creating random yitter for the x axis relative to the drill of interest so it doesn't overlap 
    #in the scatterpie plot
    x = case_when(
    drill == "Slow" ~ runif(nrow(.), 0, 1),
    drill == "Linear10" ~ runif(nrow(.), 1.1, 2),
    drill == "Linear15" ~ runif(nrow(.), 2.1, 3),
    drill == "Linear16" ~ runif(nrow(.), 3.1, 4),
    drill == "Linear18" ~ runif(nrow(.), 4.1, 5),
    drill == "NonLinearSlow" ~ runif(nrow(.), 5.1, 6),
    drill == "NonLinearFast" ~ runif(nrow(.), 6.1, 7),
    drill == "NonLinearHalf" ~ runif(nrow(.), 7.1, 8)
  )) %>% 
  group_by(athlete) %>% 
  mutate(y = case_when(
    drill == "Slow" ~ Walking/100,
    drill == "Linear10" ~ Running/100,
    drill == "Linear15" ~ Running/100,
    drill == "Linear16" ~ Running/100,
    drill == "Linear18" ~ Running/100,
    drill == "NonLinearSlow" ~ Dynamic/100,
    drill == "NonLinearFast" ~ Dynamic/100,
    drill == "NonLinearHalf" ~ Dynamic/100)) %>% 
  ungroup() %>% 
  janitor::clean_names()


```

```{r}
ggplot() +
  scatterpie::geom_scatterpie(aes(x = x, y = y, group = athlete), data = t, 
                              cols = c("dynamic","running","walking","no_locomotion"), 
                              pie_scale = 0.5) +
  geom_vline(xintercept = c(0,1,2,3,4,5,6,7,8), linetype = "dashed", color = "#838383", size = 2, alpha = 0.75) + 
  annotate("text", x = 0.5, y = 0, label = "5.4 km/h", size = 6) +
  annotate("text", x = 1.5, y = 0, label = "12 km/h", size = 6) +
  annotate("text", x = 2.5, y = 0, label = "18 km/h", size = 6) +
  annotate("text", x = 3.5, y = 0, label = "19.8 km/h", size = 6) +
  annotate("text", x = 4.5, y = 0, label = "21.6 km/h", size = 6) +
  annotate("text", x = 5.5, y = 0, label = "3 sec per cone", size = 6) +
  annotate("text", x = 6.5, y = 0, label = "All out", size = 6) +
  annotate("text", x = 7.5, y = 0, label = "All out - Half", size = 6) +
  coord_equal() + 
  scale_fill_manual(name = "Locomotion type",
                    values=c("no_locomotion"="#313131", "walking"="#0f3698", 
                             "dynamic"="#94281b", "running"="#3c8001"),
                    labels = c("No Locomotion", "Slow Locomotion", "Non-linear Locomotion", "Linear Locomotion")) + 
  #scale_x_continuous(breaks = seq(0, 7, 1), 
                     #labels = c("5.4 km/h","12 km/h","18 km/h","19.8 km/h","21.6 km/h",
                                #"3 sec per cone","all out","all out - half")) + 
  scale_y_continuous(breaks = seq(0,100,10) / 100,
                     labels = c("0","10","20","30","40","50","60","70","80","90","100")) +
  theme(legend.position = "top",
        text = element_text(size = 30),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank()) + 
  xlab("Controlled Drill") + 
  ylab("Correct prediction [%]")
```
